% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils_notExported.R
\name{bs}
\alias{bs}
\title{Return indexes for an exact number of blocks for a raster}
\usage{
bs(r, blocks)
}
\arguments{
\item{r}{a raster}

\item{blocks}{Number of desired blocks (columns, rows)}
}
\description{
Return indexes for an exact number of blocks for a raster
}
\keyword{internal}
