% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigDB.R
\name{configDB}
\alias{configDB}
\title{configDB}
\usage{
configDB(host, port, user, password, dbname)
}
\arguments{
\item{host}{(character) - host of server}

\item{port}{(character) - port of server}

\item{user}{(character) - database user}

\item{password}{(character) - database password}

\item{dbname}{(character) - database name}
}
\value{
Object with Information for connections
}
\description{
Function responsible for configuring the global class that has the database information
}
\examples{
configDB("localhost", "5432", "postgres", "", "db_accidents")
configDB("localhost", "5432", "postgres", "1234", "db_accidents")

\dontrun{
configDB("", 5432, senha, "db_accidents")
configDB(local, 5432, postgres, 123, "db_accidents")
}

}
