% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigDB.R
\name{dataTest}
\alias{dataTest}
\title{tb_dados (table of data for database test)}
\source{
http://www.testdata.com.br
}
\description{
usable this file(rPostigis/inst/extdata/tb_dados) in your database, for test package
}
\references{
acciddence of São Paulo (site)
}
