% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_gps_data}
\alias{db_gps_data}
\title{GPS tracking core data}
\format{A list containing five data frames corresponding to five GPS
sensors
\describe{
  \item{GSM01438}{data frame for sensor 01438}
  \item{GSM01508}{data frame for sensor 01508}
  \item{GSM01511}{data frame for sensor 01511}
  \item{GSM01512}{data frame for sensor 01512}
  \item{GSM02927}{data frame for sensor 02927}
}}
\source{
Urbano, Ferdinando, and Francesca Cagnacci, eds.
\emph{Spatial database for GPS wildlife tracking data: a practical guide to creating a data management system with PostgreSQL/PostGIS and R.}
Cham, Switzerland: Springer, 2014. \url{http://www.springer.com/us/book/9783319037424}.
}
\usage{
db_gps_data
}
\description{
Dataset containing a set of five raw GPS data tables from sensors
attached to roe deer in Trentino Region, Italy
}
\examples{
 data("db_gps_data")
 head(db_gps_data$GSM01438)

}
\keyword{datasets}

