% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_raster}
\alias{db_raster}
\title{Raster spatial datasets}
\format{A list containing two RasterLayer datasets
\describe{
  \item{corine06}{RasterLayer depicting land cover classification in the study area}
  \item{srtm_dem}{RasterLayer digital elevation model in the study area}
}}
\source{
Urbano, Ferdinando, and Francesca Cagnacci, eds.
\emph{Spatial database for GPS wildlife tracking data: a practical guide to creating a data management system with PostgreSQL/PostGIS and R.}
Cham, Switzerland: Springer, 2014. \url{http://www.springer.com/us/book/9783319037424}.
}
\usage{
db_raster
}
\description{
Dataset containin ancilliary raster spatial data related to GPS tracking 
of roe deer in Trentino Region, Italy
}
\examples{
 data("db_raster")
 if (require(raster, quietly = TRUE)) plot(db_raster$srtm_dem)

}
\keyword{datasets}

