% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_sensors_animals_tables}
\alias{db_sensors_animals_tables}
\title{GPS sensor and animal metadata}
\format{A list containing three data frames
\describe{
  \item{animals}{data frame containing basic information on animals}
  \item{gps_sensors}{data frame containing basic information on GPS sensors}
  \item{gps_sensors_animals}{data frame containing information on deployment of GPS sensors on animals}
}}
\source{
Urbano, Ferdinando, and Francesca Cagnacci, eds.
\emph{Spatial database for GPS wildlife tracking data: a practical guide to creating a data management system with PostgreSQL/PostGIS and R.}
Cham, Switzerland: Springer, 2014. \url{http://www.springer.com/us/book/9783319037424}.
}
\usage{
db_sensors_animals_tables
}
\description{
Dataset containing information on individual animals, sensors,
and sensors deployment on roe deer in Trentino Region, Italy
}
\examples{
 data("db_sensors_animals_tables")
 db_sensors_animals_tables$animals

}
\keyword{datasets}

