% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_vector_geom}
\alias{db_vector_geom}
\title{Vector spatial datasets}
\format{A list containing four Spatial*DataFrames
\describe{
  \item{study_area}{SpatialPolygonsDataFrame containing boundary of study area}
  \item{adm_boundaries}{SpatialPolygonsDataFrame containing administrative boundaries in study area}
  \item{meteo_stations}{SpatialPointsDataFrame containing locations of weather stations in study area}
  \item{roads}{SpatialLinesDataFrame containing representation of roads for study area}
}}
\source{
Urbano, Ferdinando, and Francesca Cagnacci, eds.
\emph{Spatial database for GPS wildlife tracking data: a practical guide to creating a data management system with PostgreSQL/PostGIS and R.}
Cham, Switzerland: Springer, 2014. \url{http://www.springer.com/us/book/9783319037424}.
}
\usage{
db_vector_geom
}
\description{
Dataset containg ancilliary vector spatial data related to GPS tracking
of roe deer in Trentino Region, Italy
}
\examples{
 data("db_vector_geom")
 if (require(sp, quietly = TRUE)) {
 plot(db_vector_geom$adm_boundaries)
 plot(db_vector_geom$roads, col = 'red', add = TRUE)
 }

}
\keyword{datasets}

