% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils_notExported.R
\name{dbBuildTableQuery}
\alias{dbBuildTableQuery}
\title{Builds CREATE TABLE query for a data frame object.}
\usage{
dbBuildTableQuery(conn = NULL, name, obj, field.types = NULL,
  row.names = FALSE)
}
\arguments{
\item{conn}{A PostgreSQL connection}

\item{name}{Table name string, length 1-2.}

\item{obj}{A data frame object.}

\item{field.types}{optional named list of the types for each field in \code{obj}}

\item{row.names}{logical, should row.name of \code{obj} be exported as a row_names field? Default is FALSE}
}
\description{
Builds CREATE TABLE query for a data frame object.
}
\note{
Adapted from RPostgreSQL::postgresqlBuildTableDefinition
}
\keyword{internal}
