% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils_notExported.R
\name{dbConnCheck}
\alias{dbConnCheck}
\title{Check if a supported PostgreSQL connection}
\usage{
dbConnCheck(conn)
}
\arguments{
\item{conn}{A PostgreSQL connection}
}
\description{
Check if a supported PostgreSQL connection
}
\keyword{internal}
