% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{dbExistsTable}
\alias{dbExistsTable}
\title{Check if a PostgreSQL table exists}
\usage{
dbExistsTable(conn, name)
}
\arguments{
\item{conn}{A PostgreSQL connection}

\item{name}{Table name string, length 1-2.}
}
\description{
Check if a PostgreSQL table exists
}
\note{
Adapted from RPostgreSQL::dbExistsTable
}
\keyword{internal}

