% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils_notExported.R
\name{dbGetDefs}
\alias{dbGetDefs}
\title{Get definitions for data frame mode reading}
\usage{
dbGetDefs(conn, name)
}
\arguments{
\item{conn}{A PostgreSQL connection}

\item{name}{Table/view name string, length 1-2.}
}
\description{
Get definitions for data frame mode reading
}
\keyword{internal}
