% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbReadSpatial.R
\name{dbReadSpatial}
\alias{dbReadSpatial}
\title{Read spatial data from a PostGIS table}
\usage{
dbReadSpatial(con, schemaname = "public", tablename, geomcol = "geom",
  idcol = NULL)
}
\arguments{
\item{con}{A valid PostgreSQL connection (from RPostgreSQL)}

\item{schemaname}{Schema where the specified table of interest resides}

\item{tablename}{Name of the database table. must contain a single
geometry column}

\item{geomcol}{Name of the geometry column in the specified table (target
table may not have more than one geometry column!) defaults to the
standard 'geom'}

\item{idcol}{Name of the column with unique IDs. This usually corresponds
 to the specified primary key.}
}
\value{
A Spatial Data Frame object
}
\description{
\code{dbReadSpatial} returns a Spatial*DataFrame
}
\details{
This is a function for reading spatial data directly from a
PostgreSQL/PostGIS table and returning the appropriate Spatial*DataFrame.
The function relies on the \code{ST_AsGeoJSON} function with PostGIS to
return the spatial data as a GeoJSON object. The non-geometry columns in
the table are read via the \code{RPostgreSQL} package.
}
\examples{
\dontrun{

}
}

