% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTableNameFix.R
\name{dbTableNameFix}
\alias{dbTableNameFix}
\title{Format input for database schema/table names.}
\usage{
dbTableNameFix(t.nm)
}
\arguments{
\item{t.nm}{Table name string, length 1-2.}
}
\value{
character vector of length 2. Each character element is in
    (escaped) double-quotes.
}
\description{
Internal rpostgis function to return common (length = 2) schema
and table name vector from various table and schema + table name
inputs.
}
\keyword{internal}

