% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{dbVersion}
\alias{dbVersion}
\title{Returns major.minor version of PostgreSQL (for version checking)}
\usage{
dbVersion(conn)
}
\arguments{
\item{conn}{A PostgreSQL connection}
}
\value{
character vector of length 2 of major,minor version.
}
\description{
Returns major.minor version of PostgreSQL (for version checking)
}
\keyword{internal}

