% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbWriteSpatial.R
\name{dbWriteSpatial}
\alias{dbWriteSpatial}
\title{Write spatial data to a PostGIS table}
\usage{
dbWriteSpatial(con, spatial.df, schemaname = "public", tablename,
  srid = NULL, replace = FALSE)
}
\arguments{
\item{con}{A valid PostgreSQL connection (from RPostgreSQL)}

\item{spatial.df}{A valid Spatial*DataFrame}

\item{schemaname}{Schema where the target table of interest will reside}

\item{tablename}{Name of the target database table.}

\item{srid}{SRID value to be passed to PostGIS (may not be needed)}

\item{replace}{Specify whether an existing table should be replaced}
}
\value{
TRUE
}
\description{
\code{dbWriteSpatial} writes a Spatial*DataFrame to a PostGIS table
}
\details{
This is a function for writing spatial data directly to
PostgreSQL/PostGIS table from an appropriate Spatial*DataFrame.
The function relies on the \code{GeoJSON} format to
transfer the spatial data. The non-geometry columns in
the table are written via the \code{RPostgreSQL} package.
}
\examples{
\dontrun{

}
}

