% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoDB.R
\docType{class}
\name{InfoDB-class}
\alias{InfoDB-class}
\alias{InfoDB}
\title{InfoDB}
\description{
Class responsável por guardar os dados para conexão com o banco de dados.
Essa é utilizada pelo metodo readDB, onde ele busca as informações dessa classe
e executa leituras no banco de dados
}
\section{Slots}{

\describe{
\item{\code{title}}{title of database - default = ''}

\item{\code{accessDriver}}{server accesses for database postgres - default = 'PostgreSQL'}

\item{\code{path}}{path your server- default = ''}

\item{\code{host}}{url the server - default = 'localhost'}

\item{\code{port}}{port the database server - default = '5432'}

\item{\code{user}}{database user - default = 'postgres'}

\item{\code{password}}{database password - default = 'root'}

\item{\code{db}}{database name - default = ''}

\item{\code{timeout}}{time limit - default = '4'}

\item{\code{encoding}}{- default = 'CP1252'}

\item{\code{dbtype}}{database type - default = 'POSTGIS'}
}}

