% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leinfo.R
\name{leinfo}
\alias{leinfo}
\title{leinfo}
\usage{
leinfo(nome, local, tempo, tipo)
}
\arguments{
\item{\code{nome}}{(character) - user name}

\item{\code{local}}{(list character) - UF desired}

\item{\code{tempo}}{(int) - amount of previous days}

\item{\code{tipos}}{(character) - vehicle}
}
\description{
function responsible for validating informations,
search the results in the database and
plot one data graphic (bar) in the window
}
\examples{
leinfo("Carlos Noronha", "SP", 2, "carros")
name <- "Carlos"
days <- 2
leinfo(name, "SP", days, "motos")

\dontrun{
leinfo("Carlos Noronha", "SP", 2)
leinfo("Carlos Noronha", "SP", "carros")
leinfo("Carlos Noronha", 2, "carros")
leinfo("SP", 2, "carros")
}

}
