% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgColumnInfo.R
\name{pgColumnInfo}
\alias{pgColumnInfo}
\title{Get information about columns in a PostgreSQL table.}
\usage{
pgColumnInfo(conn, name, allinfo = FALSE)
}
\arguments{
\item{conn}{A connection object to a PostgreSQL database}

\item{name}{A character string specifying a PostgreSQL schema (if necessary), 
and table or view name 
geometry (e.g., name = c("schema","table"))}

\item{allinfo}{logical, Get all information on table? Default is column names, types, nullable, and maximum length of character columns}
}
\value{
data frame
}
\description{
Get information about columns in a PostgreSQL table.
}
\examples{
\dontrun{

library(RPostgreSQL)
drv<-dbDriver("PostgreSQL")
conn<-dbConnect(drv,dbname='dbname',host='host',port='5432',
               user='user',password='password')
pgColumnInfo(conn,c("schema","table"))
}
}
\author{
David Bucklin \email{david.bucklin@gmail.com}
}

