% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgDrop.R
\name{pgDrop}
\alias{pgDrop}
\title{Drop table/view/schema}
\usage{
pgDrop(conn, name, type = c("table", "view", "schema"), ifexists = FALSE,
  cascade = FALSE, display = TRUE, exec = TRUE)
}
\arguments{
\item{conn}{A connection object.}

\item{name}{A character string specifying a PostgreSQL table, view
or schema name.}

\item{type}{The type of the object to comment, either \code{table}
or \code{view}}

\item{ifexists}{Do not throw an error if the table does not
exist. A notice is issued in this case.}

\item{cascade}{Automatically drop objects that depend on the table
(such as views).}

\item{display}{Logical. Whether to display the query (defaults to
\code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to
\code{TRUE}).}
}
\description{
Drop a table, a view or a schema.
}
\examples{
pgDrop(name = c("fla", "bli"), type = "view", exec = FALSE)
pgDrop(name = "fla", type = "schema", cascade = "TRUE", exec = FALSE)
}
\author{
Mathieu Basille \email{basille@ufl.edu}
}
\seealso{
The PostgreSQL documentation:
\url{http://www.postgresql.org/docs/current/static/sql-droptable.html},
\url{http://www.postgresql.org/docs/current/static/sql-dropview.html},
\url{http://www.postgresql.org/docs/current/static/sql-dropschema.html}
}

