% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgGetGeom.R
\name{pgGetGeom}
\alias{pgGetGeom}
\alias{pgGetGeomQ}
\title{Load a PostGIS geometry from a PostgreSQL table/view/query into R.}
\usage{
pgGetGeom(conn, name, geom = "geom", gid = NULL, other.cols = TRUE,
  clauses = NULL)

pgGetGeomQ(conn, query, create.view = NULL, ...)
}
\arguments{
\item{conn}{A connection object to a PostgreSQL database}

\item{name}{A character string specifying a PostgreSQL schema and
table/view name holding the geometry (e.g., \code{name =
c("schema","table")})}

\item{geom}{The name of the geometry column. (Default = \code{"geom"})}

\item{gid}{Name of the column in \code{name} holding the IDs. Should be
unique if additional columns of unique data are being
appended. \code{gid=NULL} (default) automatically creates a
new unique ID for each row in the \code{sp} object.}

\item{other.cols}{Names of specific columns in the table to
retrieve, in a character vector (e.g. \code{other.cols=c("col1","col2")}.)
The default (\code{other.cols = TRUE}) is to attach
all columns in a Spatial*DataFrame. Setting
\code{other.cols=FALSE} will return a Spatial-only object (no
data frame).}

\item{clauses}{character, additional SQL to append to modify select
query from table. Must begin with an SQL clause (e.g., "WHERE ...",
"ORDER BY ...", "LIMIT ..."); see below for examples.}

\item{query}{For \code{pgGetGeomQ}, a full SQL query including a geometry column.}

\item{create.view}{For \code{pgGetGeomQ}, optional character string specifying
a PostgreSQL schema and view name (e.g., \code{name = c("schema","view")}) 
to save the query as. If NULL, a temporary view ".rpostgis_TEMPview" is used.}

\item{...}{For \code{pgGetGeomQ}, other arguments as in \code{pgGetGeom}}
}
\value{
sp-class (SpatialPoints*, SpatialMultiPoints*, SpatialLines*, or SpatialPolygons*)
}
\description{
Retrieve point, linestring, or polygon geometries from a PostGIS
table/view/query, and convert it to an R \code{sp} object (\code{Spatial*} or
\code{Spatial*DataFrame}).
}
\details{
The query version \code{pgGetGeomQ} allows the user to enter a
full SQL query that returns a Geometry column, and save the query
as a new view if desired.
}
\examples{
\dontrun{
## Retrieve a Spatial*DataFrame with all data from table
## 'schema.tablename', with geometry in the column 'geom'
pgGetGeom(conn, c("schema", "tablename"))
## Return a Spatial*DataFrame with columns c1 & c2 as data
pgGetGeom(conn, c("schema", "tablename"), other.cols = c("c1","c2"))
## Return a Spatial*-only (no data frame), 
## retaining id from table as rownames
pgGetGeom(conn, c("schema", "tablename"), gid = "table_id",
  other.cols = FALSE)
## Return a Spatial*-only (no data frame), 
## retaining id from table as rownames and with a subset of the data
pgGetGeom(conn, c("schema", "roads"), geom = "roadgeom", gid = "road_ID",
    other.cols = FALSE, clauses  = "WHERE field = 'highway'")

## Query version
pgGetGeomQ(conn, "SELECT r.road_ID as id, ST_Buffer(r.roadgeom, 100) as geom 
                  FROM
                     schema.roads r,
                     schema.boundary b
                  WHERE 
                     ST_Intersects(r.roadgeom, b.boundgeom);")
}
}
\author{
David Bucklin \email{dbucklin@ufl.edu}

Mathieu Basille \email{basille@ufl.edu}
}
