% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils_notExported.R
\name{pgGetSRID}
\alias{pgGetSRID}
\title{Get SRID(s) from a geometry/geography column in a full table}
\usage{
pgGetSRID(conn, name, geom)
}
\arguments{
\item{conn}{A PostgreSQL connection}

\item{name}{A schema/table name}

\item{geom}{a geometry or geography column name}
}
\description{
Get SRID(s) from a geometry/geography column in a full table
}
\keyword{internal}
