% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgInsert.R
\name{pgInsert}
\alias{pgInsert}
\title{Inserts data from a pgInsertize* object into a PostgreSQL table}
\usage{
pgInsert(conn, name, pgi, encoding = NULL)
}
\arguments{
\item{conn}{A connection object to a PostgreSQL database}

\item{name}{character strings specifying a PostgreSQL schema and table name to insert into (e.g., name = c("schema","table"))}

\item{pgi}{A list of columns to insert to and the formatted insert data (the output object from pgInsertize() or pgInsertizeGeom())}

\item{encoding}{Character vector of length 2, containing the from/to encodings for the 
data (as in the function \code{iconv}. For example, if your dataset contain certain latin characters (e.g., accent marks),
and the database is in UTF-8, use \code{encoding = c("latin1","UTF-8")}. Left NULL, no conversion will be done.}
}
\value{
DBIResult
}
\description{
Inserts data from a pgInsertize* object into a PostgreSQL table
}
\examples{

library(sp)
data(meuse)
coords <- SpatialPoints(meuse[, c("x", "y")])
spdf<- SpatialPointsDataFrame(coords, meuse)

#remove "." from column name
colnames(spdf@data)[colnames(spdf@data) == 'dist.m']<-"dist_m"

#format data for insert
pgi<-pgInsertizeGeom(spdf,geom="point")

\dontrun{
library(RPostgreSQL)
drv<-dbDriver("PostgreSQL")
conn<-dbConnect(drv,dbname='dbname',host='host',port='5432',
               user='user',password='password')

# insert data in database table (note that an error will be given if 
# all insert columns do not match exactly to database table columns)
pgInsert(conn,c("schema","meuse_data"),pgi=pgi)
}
}
\author{
David Bucklin \email{david.bucklin@gmail.com}
}

