% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgInsertize.r
\name{pgInsertize}
\alias{pgInsertize}
\title{Formats an R data frame for insert into a PostgreSQL table (for use with pgInsert)}
\usage{
pgInsertize(df, force.match = NULL, conn = NULL)
}
\arguments{
\item{df}{A data frame}

\item{force.match}{character, schema and table of the PostgreSQL table to compare columns of data frame with 
If specified, only columns in the data frame that exactly match the database table will be kept, and reordered
to match the database table. If NULL, all columns will be kept in the same order given in the data frame.}

\item{conn}{A database connection (required if a table is given in for "force.match" parameter)}
}
\value{
List containing two character strings- (1) db.cols.insert, a character string of the database column
names to make inserts on, and (2) insert.data, a character string of the data to insert. See examples for 
usage within the \code{pgInsert} function.
}
\description{
Formats an R data frame for insert into a PostgreSQL table (for use with pgInsert)
}
\examples{

\dontrun{
#connect to database
library(RPostgreSQL)
drv<-dbDriver("PostgreSQL")
conn<-dbConnect(drv,dbname='dbname',host='host',port='5432',
               user='user',password='password')
}

data<-data.frame(a=c(1,2,3),b=c(4,NA,6),c=c(7,'text',9))

#format all columns for insert
values<-pgInsertize(df=data)

\dontrun{
# insert data in database table (note that an error will be given if all insert columns 
# do not match exactly to database table columns)
pgInsert(conn,c("schema","table"),pgi=values)

##
#run with forced matching of database table column names
values<-pgInsertize(df=data,force.match=c("schema","table"),conn=conn)

pgInsert(conn,c("schema","table"),pgi=values)
}
}
\author{
David Bucklin \email{david.bucklin@gmail.com}
}

