% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgis2spts.r
\name{pgis2spts}
\alias{pgis2spts}
\title{Load a point geometry stored in a PostgreSQL database into R.}
\usage{
pgis2spts(conn, table, geom = "geom", gid = "gid", proj = NULL,
  other.cols = "*", query = NULL)
}
\arguments{
\item{conn}{A connection object created in RPostgreSQL package.}

\item{table}{character, Name of the schema-qualified table in Postgresql holding the geometry.}

\item{geom}{character, Name of the column in 'table' holding the geometry object (Default = 'geom')}

\item{gid}{character, Name of the column in 'table' holding the ID for each point geometry. Should be unique if additional columns of unique data are being appended. (Default = 'gid')}

\item{proj}{numeric, Can be set to TRUE to automatically take the SRID for the table in the database. Alternatively, the number of EPSG-specified projection of the geometry (Default is NULL, resulting in no projection.)}

\item{other.cols}{character, names of additional columns from table (comma-seperated) to append to dataset (Default is all columns, other.cols=NULL returns a SpatialPoints object)}

\item{query}{character, additional SQL to append to modify select query from table}
}
\value{
SpatialPointsDataFrame or SpatialPoints
}
\examples{
\dontrun{
library(RPostgreSQL)
drv<-dbDriver("PostgreSQL")
conn<-dbConnect(drv,dbname='dbname',host='host',port='5432',user='user',password='password')

pgis2spts(conn,'schema.tablename')
pgis2spts(conn,'schema.cities',geom='citygeom',gid='city_ID',proj=4326,query = "AND city_population > 10000")
}
}
\author{
David Bucklin \email{david.bucklin@gmail.com}
}

