% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgtablenamefix.R
\name{pgtablenamefix}
\alias{pgtablenamefix}
\title{Format input for database schema/table names}
\usage{
pgtablenamefix(t.nm)
}
\arguments{
\item{t.nm}{Table name string, length 1-2.}
}
\value{
character vector of length 2. Each character element is in (escaped) double-quotes.
}
\description{
Internal rpostgis function to return common (length = 2) schema and table
name vector from various table and schema + table name inputs.
}
\examples{
name<-c("schema","table")
pgtablenamefix(name)

name<-"schema.table"
pgtablenamefix(name)

#default schema (public) is added to tables
name<-"table"
pgtablenamefix(name)

#schema or table names with "." need to be given in two length vectors:
name<-c("schema","ta.ble")
pgtablenamefix(name)
}

