% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadDB.R
\name{readDB}
\alias{readDB}
\title{readDB}
\usage{
readDB(objectInfo, date, type, uf)
}
\arguments{
\item{objectInfo}{(InfoDB) - Object with informations for connections}

\item{date}{(character) - amount of previous days (Y-m-d)}

\item{type}{(character) - vehicle}

\item{uf}{(int) - UF desired}
}
\value{
Number of occurrences (int)
}
\description{
function responsible for searching the database and
returning numbers of occurrences on the select date
}
