% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils_notExported.R
\name{warn_deprecated_sp}
\alias{warn_deprecated_sp}
\title{Warning for deprecated sp and raster}
\usage{
warn_deprecated_sp(data, what)
}
\arguments{
\item{data}{dataset}

\item{what}{text to show}
}
\description{
Warning for deprecated sp and raster
}
\keyword{internal}
