% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMultiOGR.R
\name{writeMultiOGR}
\alias{writeMultiOGR}
\title{Write an MultiPolygon into PostgreSQL data source.}
\usage{
writeMultiOGR(x, dsn, nln = " ", verbose = F, srid = "4326",
  dbname = "mydb", host = " ", username = " ", simplify = T,
  shp2pgsql = F)
}
\arguments{
\item{x}{a object}

\item{dsn}{PostgreSQL DSNs}

\item{nln}{a character indicating layer option}

\item{verbose}{a boolean variable indicating whether to display verbose information}

\item{srid}{a integer variable indicating spatial reference id of OGR data source.The defalut value is 4326(WGS84)}

\item{dbname}{a character indicating database name}

\item{host}{If shp2pgsql is true,a charater is required to indicate hostname or ip address of PostgreSQL Server}

\item{username}{If shp2pgsql is true,a charater is required to indicate current user}
}
\value{
Spatial DataFrame
}
\keyword{writeMultiOGR}

