% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils_notExported.R
\name{dbVersion}
\alias{dbVersion}
\title{Returns major.minor version of PostgreSQL (for version checking)}
\usage{
dbVersion(conn)
}
\arguments{
\item{conn}{A PostgreSQL connection}
}
\value{
numeric vector of length 3 of major,minor,bug version.
}
\description{
Returns major.minor version of PostgreSQL (for version checking)
}
\keyword{internal}
