% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leinfo.R
\name{leinfo}
\alias{leinfo}
\title{leinfo}
\usage{
leinfo(objectInfo, nome, local, tempo, tipo)
}
\arguments{
\item{objectInfo}{(InfoDB) - Object with informations for connections}

\item{nome}{(character) - user name}

\item{local}{(character) - UF desired}

\item{tempo}{(int) - amount of previous days}

\item{tipo}{(character) - vehicle}
}
\description{
function responsible for validating informations,
search the results in the database and
plot one data graphic (bar) in the window
}
\examples{
con <- configDB("localhost", "5432", "postgres", "1234", "db_lingR")
leinfo(con, "Carlos Noronha", "SP", 2, "carros")
name <- "Carlos"
days <- 2
leinfo(con, name, "SP", days, "motos")

\dontrun{
leinfo(con, "Carlos Noronha", "SP", 2)
leinfo("Carlos Noronha", "SP", "carros")
leinfo("host", Carlos Noronha", 2, "carros")
leinfo("con", SP", 2, "carros")
}

}
