% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils_notExported.R
\name{pgCheckGeom}
\alias{pgCheckGeom}
\title{Check if geometry or geography column exists in a table,
and return the column name for use in a query.}
\usage{
pgCheckGeom(conn, namechar, geom)
}
\arguments{
\item{conn}{A PostgreSQL connection}

\item{namechar}{A table name formatted for use in a query}

\item{geom}{a geometry or geography column name}
}
\description{
Check if geometry or geography column exists in a table,
and return the column name for use in a query.
}
\keyword{internal}
