% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgGetGeom.R
\name{pgGetLines}
\alias{pgGetLines}
\title{Load a PostGIS linestring geometry from a PostgreSQL table/view into R.}
\usage{
pgGetLines(conn, name, geom = "geom", gid = NULL, other.cols = "*",
  clauses = NULL)
}
\value{
SpatialLinesDataFrame or SpatialLines
}
\description{
Load a PostGIS linestring geometry from a PostgreSQL table/view into R.
}
\examples{
\dontrun{
pgGetLines(conn, c("schema", "tablename"))
pgGetLines(conn, c("schema", "roads"), geom = "roadgeom", gid = "road_ID",
    other.cols = NULL, clauses  = "WHERE field = 'highway'")
}
}
\author{
David Bucklin \email{david.bucklin@gmail.com}

Mathieu Basille \email{basille@ufl.edu}
}
\keyword{internal}
