% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgInsert.R
\name{pgInsert}
\alias{pgInsert}
\title{Inserts data from a pgInsertize* object into a PostgreSQL table}
\usage{
pgInsert(conn, pgi, name = NULL, encoding = NULL)
}
\arguments{
\item{conn}{A connection object to a PostgreSQL database}

\item{pgi}{The output PostgreSQL insert object (pgi) created by pgInsertize() or pgInsertizeGeom())}

\item{name}{character strings specifying a PostgreSQL schema and table name to insert into (e.g., name = c("schema","table")). 
If table was specified in the pgInsertize* through create.table or force.match, leave this NULL.}

\item{encoding}{Character vector of length 2, containing the from/to encodings for the 
data (as in the function \code{iconv}). For example, if the dataset contain certain latin characters (e.g., accent marks),
and the database is in UTF-8, use \code{encoding = c("latin1","UTF-8")}. Left NULL, no conversion will be done.}
}
\value{
DBIResult
}
\description{
This function takes a \code{pgi} list output object from \code{pgInsertize} or \code{pgInsertizeGeom} and 
performs the database insert (and table creation, if specified in the previous functions) on
the database. If \code{create.table} or \code{force.match} were not specified in the \code{pgInsertize*}
statement, the table to insert into should be specified in \code{name} in this function.
If a new table is created but the data insert statement fails, the new table is dropped from the database (a
message will be given).
}
\examples{

library(sp)
data(meuse)
coords <- SpatialPoints(meuse[, c("x", "y")])
spdf<- SpatialPointsDataFrame(coords, meuse)

#format data for insert
pgi<-pgInsertizeGeom(spdf,geom="point")

\dontrun{
library(RPostgreSQL)
drv<-dbDriver("PostgreSQL")
conn<-dbConnect(drv,dbname='dbname',host='host',port='5432',
               user='user',password='password')

# insert data in database table (note that an error will be given if 
# all insert columns do not match exactly to database table columns)
pgInsert(conn,pgi=pgi,name=c("schema","meuse_data"))
}
}
\author{
David Bucklin \email{david.bucklin@gmail.com}
}

