% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgInsertize.r
\name{pgInsertize}
\alias{pgInsertize}
\title{Formats an R data frame as a character string for use in a PostgreSQL insert statement.}
\usage{
pgInsertize(df, db.na = "NULL")
}
\arguments{
\item{df}{A data frame}

\item{db.na}{A character string, value to change NAs to (defaults to "NULL")}
}
\value{
Character string which is suitable for pasting into a SQL INSERT statement
}
\description{
Formats an R data frame as a character string as insert values for a PostgreSQL insert statement.
}
\examples{
\dontrun{
library(RPostgreSQL)
drv<-dbDriver("PostgreSQL")
conn<-dbConnect(drv,dbname='dbname',host='host',port='5432',
               user='user',password='password')
}

data<-data.frame(a=c(1,2,3),b=c(4,NA,6),c=c(7,'text',9))
values<-pgInsertize(data)

\dontrun{
dbSendQuery(conn,paste0("INSERT INTO schema.table (col1,col2,col3) VALUES ",values,";"))
}
}
\author{
David Bucklin \email{david.bucklin@gmail.com}
}

