% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgInsertizeGeom.r
\name{pgInsertizeGeom}
\alias{pgInsertizeGeom}
\title{Formats an R Spatial*DataFrame as character string insert values for a PostgreSQL insert statement.}
\usage{
pgInsertizeGeom(sdf, db.na = "NULL")
}
\arguments{
\item{sdf}{A Spatial*DataFrame}

\item{db.na}{A character string, value to change NAs to (defaults to "NULL")}
}
\value{
Character string which is suitable for pasting into a SQL INSERT statement for PostGIS spatial tables.
}
\description{
Formats an R Spatial*DataFrame as character string insert values for a PostgreSQL insert statement.
}
\examples{

library(sp)
data(meuse)
coords <- SpatialPoints(meuse[, c("x", "y")])
spdf<- SpatialPointsDataFrame(coords, meuse)
values<-pgInsertizeGeom(sdf=spdf)

\dontrun{

library(RPostgreSQL)
drv<-dbDriver("PostgreSQL")
conn<-dbConnect(drv,dbname='dbname',host='host',port='5432',
               user='user',password='password')

##Note that the geometry column must always be last in the insert column list
dbSendQuery(conn,paste0("INSERT INTO schema.table (col1,col2,col3,geom) VALUES ",values,";"))
}
}
\author{
David Bucklin \email{david.bucklin@gmail.com}
}

