% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgListGeom.R
\name{pgListGeom}
\alias{pgListGeom}
\alias{pgListRast}
\title{List geometries/rasters}
\usage{
pgListGeom(conn, geog = TRUE)

pgListRast(conn)
}
\arguments{
\item{conn}{A PostgreSQL database connection.}

\item{geog}{Logical. For pgListGeom, whether to include PostGIS geography-type 
columns stored in the database}
}
\value{
If \code{exec = TRUE}, a data frame with schema, table,
    geometry/(geography) or raster (for pgListRast) column, and geometry/(geography) type.
}
\description{
List all geometry/(geography) or raster columns available in a PostGIS database.
}
\examples{
\dontrun{
pgListGeom(conn)

pgListRast(conn)
}
}
\author{
David Bucklin \email{david.bucklin@gmail.com}
}
