% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgListGeom.R
\name{pgListGeom}
\alias{pgListGeom}
\title{List geometries.}
\usage{
pgListGeom(conn, display = TRUE, exec = TRUE)
}
\arguments{
\item{conn}{A PostgreSQL database connection.}

\item{display}{Logical. Whether to display the query (defaults to
\code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to
\code{TRUE}).}
}
\value{
If \code{exec = TRUE}, a data frame with schema, table,
    geometry column, and geometry type.
}
\description{
List all geometries in a PostGIS database.
}
\examples{
\dontrun{
pgListGeom(conn)
}
}
\author{
David Bucklin \email{dbucklin@ufl.edu}
}

