% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOgrSql.R
\name{readOgrSql}
\alias{readOgrSql}
\title{Define a filename from current date.}
\usage{
readOgrSql(dsn, sql, ...)
}
\arguments{
\item{dsn}{PostgreSQL DSNs}
}
\value{
Spatial DataFrame
}
\description{
Define a filename from current date.
}
\examples{
\dontrun{ strSQL = "
SELECT gid, name,
ST_Transform(geom, 4326)::geometry(MultiPolygon, 4326) AS geom
FROM geo_states"
spdfStates = readOgrSql(dsn, strSQL, stringsAsFactors=FALSE)}

}
\references{
\url{https://geospatial.commons.gc.cuny.edu/2013/12/31/subsetting-in-readogr/}
}
\keyword{readOgrSql}

