\name{RpostGIS-package}
\alias{RpostGIS-package}
\alias{RpostGIS}
\docType{package}
\title{Wrapper to PostGIS functions}
\description{
Provides bindings to PostGIS functionalities. Through this bindings (due to OGC specifications)  numerous spatial functions can be performed by simple SQL statements.
}

\details{
\tabular{ll}{
Package: \tab RpostGIS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-10-07\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Norbert Solymosi

Maintainer: Norbert Solymosi <solymosi.norbert@gmail.com>
}
\references{
http://postgis.net//\cr
http://www.opengeospatial.org/standards\cr

}
\keyword{ spatial }
\keyword{ SQL }

