% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rpostgis.R
\docType{package}
\name{Rpostgis}
\alias{Rpostgis}
\alias{Rpostgis-package}
\title{Rpostgis: A package for import/export of spatial data from PostGIS}
\description{
Rpostgis provides simple access to PostgreSQL/PostGIS spatial databases
without the need to build the rgdal packages from source (pre-built binaries
on CRAN are not built against the PostgreSQL/PostGIS libraries). Initially,
the focus of this package is on reading and writing data, but further
development for various spatial functions could be considered.
}
\section{Rpostgis functions}{

  \itemize{
    \item dbReadSpatial
    \item dbWriteSpatial
  }
}

